refactor: :common 과 :ui-components 합체
