--- common/src/main/AndroidManifest.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ Designed and developed by 2022 SungbinLand, Team Duckie
-  ~
-  ~ Licensed under the MIT.
-  ~ Please see full license: https://github.com/sungbinland/quack-quack/blob/main/LICENSE
-  -->
-
-<manifest />
