+++ ui-components/src/main/kotlin/team/duckie/quackquack/ui/util/detekt.kt
@@ -5,7 +5,7 @@
  * Please see full license: https://github.com/sungbinland/quack-quack/blob/main/LICENSE
  */
 
-package team.duckie.quackquack.common
+package team.duckie.quackquack.ui.util
 
 /**
  * MagicNumber 을 예외적으로 허용할 때 사용하는 어노테이션 입니다.
@@ -13,4 +13,4 @@ package team.duckie.quackquack.common
  * @param because 이 경우에 MagicNumber 을 왜 허용하는지 명시해야 합니다.
  */
 @Retention(AnnotationRetention.SOURCE)
-annotation class AllowMagicNumber(@Suppress("unused") val because: String)
+internal annotation class AllowMagicNumber(@Suppress("unused") val because: String)
