+++ ui-components/src/main/kotlin/team/duckie/quackquack/ui/util/dsl.kt
@@ -5,10 +5,10 @@
  * Please see full license: https://github.com/sungbinland/quack-quack/blob/main/LICENSE
  */
 
-package team.duckie.quackquack.common
+package team.duckie.quackquack.ui.util
 
 /**
  * QuackQuack 에서 DSL Builder 를 만들 때 사용됩니다.
  */
 @DslMarker
-annotation class QuackDsl
+internal annotation class QuackDsl
